requirements.txt
streamlit>=1.23.0
pandas>=1.5.3
numpy>=1.24.3
plotly>=5.14.0
openpyxl>=3.1.2
xlsxwriter>=3.1.0
google-cloud-storage>=2.7.0